renamed getCommentLocation to getThreadLocation, fixed regex to reject properly
